+++ src/main/java/com/ybe/tr1ll1on/domain/product/dto/response/ProductFacilityResponse.java
@@ -1,4 +1,4 @@
-package com.ybe.tr1ll1on.domain.product.response;
+package com.ybe.tr1ll1on.domain.product.dto.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
